# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StrategistOS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DiagnosisResultRoiActionsItem = {
  action: string;
  impact: string;
  timeframe: string;
};

export interface DiagnosisResult {
  strategicDiagnosis: string;
  leverageScore: number;
  bottleneckAnalysis: string;
  opportunityMap: string[];
  riskMap: string[];
  roiActions: DiagnosisResultRoiActionsItem[];
  eliteOperatorNextStep: string;
}

export interface Session {
  id: number;
  title: string;
  goal: string;
  industry: string;
  leverageScore?: number;
  createdAt: string;
  updatedAt: string;
  diagnosisResult?: DiagnosisResult | null;
}

export interface CreateSessionBody {
  title: string;
  goal: string;
  industry: string;
}

export type SessionsSummaryTopIndustriesItem = {
  industry: string;
  count: number;
};

export interface SessionsSummary {
  totalSessions: number;
  averageLeverageScore: number;
  recentSessions: Session[];
  topIndustries: SessionsSummaryTopIndustriesItem[];
}

export interface DiagnosisInput {
  goal: string;
  industry: string;
  assets: string;
  constraints: string;
  deadline: string;
  desiredOutcome: string;
  bottleneck: string;
  sessionId?: number | null;
}

export interface ScorecardInput {
  goal: string;
  industry: string;
  currentStatus: string;
  sessionId?: number | null;
}

export interface ScorecardDimension {
  name: string;
  score: number;
  reasoning: string;
  howToImprove: string;
  eliteRecommendation: string;
}

export interface ScorecardResult {
  overallScore: number;
  dimensions: ScorecardDimension[];
  topPriority: string;
  strategicSummary: string;
}

export type PromptGenerateInputCategory =
  (typeof PromptGenerateInputCategory)[keyof typeof PromptGenerateInputCategory];

export const PromptGenerateInputCategory = {
  "business-strategy": "business-strategy",
  "academic-writing": "academic-writing",
  "ai-automation": "ai-automation",
  "content-creation": "content-creation",
  "linkedin-positioning": "linkedin-positioning",
  "offer-creation": "offer-creation",
  "replit-app-building": "replit-app-building",
  "data-science": "data-science",
  "phd-research": "phd-research",
} as const;

export interface PromptGenerateInput {
  category: PromptGenerateInputCategory;
  context: string;
  goal: string;
}

export interface GeneratedPrompt {
  title: string;
  category: string;
  prompt: string;
  usage: string;
  variables: string[];
}

export interface SavePromptBody {
  title: string;
  category: string;
  prompt: string;
  usage: string;
  variables: string[];
}

export interface SavedPrompt {
  id: number;
  title: string;
  category: string;
  prompt: string;
  usage: string;
  variables: string[];
  createdAt: string;
}

export interface OpportunityInput {
  skills: string;
  experience: string;
  tools: string;
  projects: string;
  targetAudience: string;
  desiredPath: string;
}

export type OpportunityResultRoadmap30DayItem = {
  week: number;
  focus: string;
  actions: string[];
};

export interface OpportunityResult {
  positioningAngle: string;
  bestNiche: string;
  offerIdea: string;
  contentAngle: string;
  proofAsset: string;
  roadmap30Day: OpportunityResultRoadmap30DayItem[];
}

export interface AiWorkflow {
  id: number;
  name: string;
  description: string;
  trigger: string;
  inputs: string[];
  aiTask: string;
  tools: string[];
  output: string;
  humanReviewStep: string;
  monetisationUseCase: string;
  isTemplate: boolean;
  createdAt: string;
}

export interface CreateWorkflowBody {
  name: string;
  description: string;
  trigger: string;
  inputs: string[];
  aiTask: string;
  tools: string[];
  output: string;
  humanReviewStep: string;
  monetisationUseCase: string;
}

export interface WorkflowTemplate {
  id: number;
  name: string;
  description: string;
  trigger: string;
  inputs: string[];
  aiTask: string;
  tools: string[];
  output: string;
  humanReviewStep: string;
  monetisationUseCase: string;
  isTemplate: boolean;
  createdAt: string;
}

export interface PlannerInput {
  strategicGoal: string;
  keyRecommendation: string;
  resources: string;
  constraints: string;
}

export type DayActionPriority =
  (typeof DayActionPriority)[keyof typeof DayActionPriority];

export const DayActionPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface DayAction {
  day: number;
  task: string;
  priority: DayActionPriority;
  estimatedTime: string;
}

export type ExecutionPlanThirtyDayRoadmapItem = {
  week: number;
  theme: string;
  milestones: string[];
};

export type ExecutionPlanRisksItem = {
  risk: string;
  mitigation: string;
};

export interface ExecutionPlan {
  title: string;
  sevenDaySprint: DayAction[];
  thirtyDayRoadmap: ExecutionPlanThirtyDayRoadmapItem[];
  successMetrics: string[];
  risks: ExecutionPlanRisksItem[];
  reviewQuestions: string[];
}

export interface SavePlanBody {
  title: string;
  plan: ExecutionPlan;
}

export interface SavedPlan {
  id: number;
  title: string;
  plan: ExecutionPlan;
  createdAt: string;
}

export type PortfolioSystemsItem = {
  title: string;
  description: string;
  outcome: string;
};

export type PortfolioCaseStudiesItem = {
  title: string;
  challenge: string;
  approach: string;
  result: string;
  industry: string;
};

export interface Portfolio {
  name: string;
  tagline: string;
  positioningStatement: string;
  philosophy: string;
  contactEmail: string;
  bookingUrl?: string;
  systems: PortfolioSystemsItem[];
  caseStudies: PortfolioCaseStudiesItem[];
}

export type UpdatePortfolioBodySystemsItem = {
  title: string;
  description: string;
  outcome: string;
};

export type UpdatePortfolioBodyCaseStudiesItem = {
  title: string;
  challenge: string;
  approach: string;
  result: string;
  industry: string;
};

export interface UpdatePortfolioBody {
  name?: string;
  tagline?: string;
  positioningStatement?: string;
  philosophy?: string;
  contactEmail?: string;
  bookingUrl?: string;
  systems?: UpdatePortfolioBodySystemsItem[];
  caseStudies?: UpdatePortfolioBodyCaseStudiesItem[];
}