# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StrategistOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prompts/generate:
    post:
      operationId: generatePrompt
      tags: [prompts]
      summary: Generate a premium prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptGenerateInput"
      responses:
        "200":
          description: Generated prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedPrompt"
  /prompts/saved:
    get:
      operationId: listSavedPrompts
      tags: [prompts]
      summary: List saved prompts
      responses:
        "200":
          description: List of saved prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedPrompt"
    post:
      operationId: savePrompt
      tags: [prompts]
      summary: Save a generated prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePromptBody"
      responses:
        "201":
          description: Saved prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPrompt"
  /prompts/saved/{id}:
    delete:
      operationId: deleteSavedPrompt
      tags: [prompts]
      summary: Delete a saved prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List saved AI workflows
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiWorkflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a new AI workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "201":
          description: Created workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiWorkflow"
...
  /workflows/{id}:
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /planner:
    post:
      operationId: createExecutionPlan
      tags: [planner]
      summary: Create an elite execution plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerInput"
      responses:
        "200":
          description: Execution plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlan"
  /planner/saved:
    get:
      operationId: listSavedPlans
      tags: [planner]
      summary: List saved execution plans
      responses:
        "200":
          description: List of saved plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedPlan"
    post:
      operationId: savePlan
      tags: [planner]
      summary: Save an execution plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePlanBody"
      responses:
        "201":
          description: Saved plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPlan"
...
components:
  schemas:
...
    PromptGenerateInput:
      type: object
      properties:
        category:
          type: string
          enum:
            - business-strategy
            - academic-writing
            - ai-automation
            - content-creation
            - linkedin-positioning
            - offer-creation
            - replit-app-building
            - data-science
            - phd-research
        context:
          type: string
        goal:
          type: string
      required:
        - category
        - context
        - goal
    GeneratedPrompt:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        prompt:
          type: string
        usage:
          type: string
        variables:
          type: array
          items:
            type: string
      required:
        - title
        - category
        - prompt
        - usage
        - variables
    SavePromptBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        prompt:
          type: string
        usage:
          type: string
        variables:
          type: array
          items:
            type: string
      required:
        - title
        - category
        - prompt
        - usage
        - variables
    SavedPrompt:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        prompt:
          type: string
        usage:
          type: string
        variables:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - prompt
        - usage
        - variables
        - createdAt
...
    AiWorkflow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        trigger:
          type: string
        inputs:
          type: array
          items:
            type: string
        aiTask:
          type: string
        tools:
          type: array
          items:
            type: string
        output:
          type: string
        humanReviewStep:
          type: string
        monetisationUseCase:
          type: string
        isTemplate:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - trigger
        - inputs
        - aiTask
        - tools
        - output
        - humanReviewStep
        - monetisationUseCase
        - isTemplate
        - createdAt
    CreateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        trigger:
          type: string
        inputs:
          type: array
          items:
            type: string
        aiTask:
          type: string
        tools:
          type: array
          items:
            type: string
        output:
          type: string
        humanReviewStep:
          type: string
        monetisationUseCase:
          type: string
      required:
        - name
        - description
        - trigger
        - inputs
        - aiTask
        - tools
        - output
        - humanReviewStep
        - monetisationUseCase
...
    PlannerInput:
      type: object
      properties:
        strategicGoal:
          type: string
        keyRecommendation:
          type: string
        resources:
          type: string
        constraints:
          type: string
      required:
        - strategicGoal
        - keyRecommendation
        - resources
        - constraints
    DayAction:
      type: object
      properties:
        day:
          type: integer
        task:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        estimatedTime:
          type: string
      required:
        - day
        - task
        - priority
        - estimatedTime
    ExecutionPlan:
      type: object
      properties:
        title:
          type: string
        sevenDaySprint:
          type: array
          items:
            $ref: "#/components/schemas/DayAction"
        thirtyDayRoadmap:
          type: array
          items:
            type: object
            properties:
              week:
                type: integer
              theme:
                type: string
              milestones:
                type: array
                items:
                  type: string
            required:
              - week
              - theme
              - milestones
        successMetrics:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: object
            properties:
              risk:
                type: string
              mitigation:
                type: string
            required:
              - risk
              - mitigation
        reviewQuestions:
          type: array
          items:
            type: string
      required:
        - title
        - sevenDaySprint
        - thirtyDayRoadmap
        - successMetrics
        - risks
        - reviewQuestions
    SavePlanBody:
      type: object
      properties:
        title:
          type: string
        plan:
          $ref: "#/components/schemas/ExecutionPlan"
      required:
        - title
        - plan
    SavedPlan:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        plan:
          $ref: "#/components/schemas/ExecutionPlan"
        createdAt:
          type: string
      required:
        - id
        - title
        - plan
        - createdAt
...